examples/python/requirements.txt
google-adk>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pytest>=7.0.0
